Moving opentelemetry imports inside
As we found, opentelemetry does extensive use of pkg_resources. Which is fine, but doesn't play well with pyinstaller

None of the pkg_resources relevant stuff is automatically pulled into pyinstaller. So we always have to add a bunch of hooks to manually add them

The problem is that there is not only that, there is an `iter_entry_points`, which https://github.com/pyinstaller/pyinstaller/issues/3050 documents as a non-straightforward thing to do

Since we are already not running opentelemetry on enterprise builds, the effort, while possible, doesn't make sense for what will result in "just running as before without the package".

So we are just moving the imports to inside the init tracing. It shouldn't affect anything on non-enteprise builds.

We are also adding hooks folders, because that part at least is solved for if/when we actuall need this stuff on enteprise