Applying mutation tests on shared
Playing around with mutation tests on shared. I think it's a good way to get our feet wet into the mutation testing world and understand what it means.

There is a lot of repetition between the mutations workflow and the testing one. But I don't want to damage the testing workflow because of mutation testing problems.

I am hoping we can later improve this with "workflows" from CircleCI: https://circleci.com/blog/persisting-data-in-workflows-when-to-use-caching-artifacts-and-workspaces/

But mutation tests is working. It only applies to lines present at the git diff (it takes way too long otherwise) and it produces interesting results which are currently saved on the "tests" tab on CircleCI: https://app.circleci.com/pipelines/github/codecov/shared/1239/workflows/82ebe8af-f5ac-4f09-ad01-8fcafd810aff/jobs/2518/tests