Making basic plan the default
Notice that this default does not go on the actual database level. The SQL for the migration is empty:

```
(apienv) ➜  api git:(master) ✗ python manage.py sqlmigrate codecov_auth 0006
/Users/thiagorramos/.pyenv/versions/apienv/lib/python3.9/site-packages/cerberus/validator.py:1607: UserWarning: No validation schema is defined for the arguments of rule 'check_aggregation_fields'
  warn(
BEGIN;
--
-- Alter field plan on owner
--
--
-- Alter field owner on ownerprofile
--
--
-- Alter field key on repositorytoken
--
--
-- Alter field valid_until on repositorytoken
--
COMMIT;
```

So this will only change the default for new users logging in for now. We will also soon change the database level so the worker imports can do it too.