Catching torngit error on apply_diff as well
On the `save_report_results` step on `UploadProcessorTask`, we fetch the diff from Github in order to calculate diff totals.

We already catch `TorngitClientError` exceptions there, because it's better to have the report and no diff than nothing.

But recently there have been a couple of cases where other exceptions occurred. I believe we should cover them as well