Removing extra strip from lua coverage
We have had Lua reports where the first few empty lines are relevant.

It seems like the problem is due to a `strip`that is not needed

We are also adding some tests. From the tests, it looks like that xmls and jsons still read fine once parsed with extra line breaks in the beginning, as this was the first assumption of what the problem was.