Implementing Static Analysis Views
As described on 67fa52b29a92212208be4ff09dfcd1e794e4bb52 (see there for more details on the decision making process), the static analysis system will have some similarities with the Git System. It will have a, per repository, a registry of file versions (identified by their checksum), and will save new file versions only when they actually change.

This will allow customer to have less things to upload on each time, and our database will more quickly be able to tell differences between two different commits.

Now we are implementing the serializers and views that will allow that behavior. We will error on the minimalist side and only implement the views we know we need for now. Later we can implement listing and some custom actions, but for now we only need the creation view.

The permission side is very similar to the `profiling` authentication. We have tokens that are limited on which repos they can serve and on each scopes (powers) they have. We will use the `SpecificScopePermission` for that.

The request payload will be:

```
{
    "commit": "1fbe7f5d509ac5721d970d9f4ff4019fcdb53797",
    "filepaths": [
        {
            "filepath": "a.py",
            "file_hash": "b29af6411ae3302da652cccd2878fb69"
        },
        {
            "filepath": "labels.py",
            "file_hash": "8d77b5c0058ad0880c99218877f2e27e"
        },
        // ...
        {
            "filepath": "tests/test_six.py",
            "file_hash": "557e32ff421eec9fedeaba35e876b10d"
        }
    ]
}
```

which essentially tells which commit the customer wants to create the analysis for, and the checksums, with 128-bits, for each commit.

And the response:

```
{
    "external_id": "58f51cdf-7621-463b-b022-fce3de0696e7",
    "commit": "1fbe7f5d509ac5721d970d9f4ff4019fcdb53797",
    "filepaths": [
        {
            "filepath": "a.py",
            "file_hash": "b29af641-1ae3-302d-a652-cccd2878fb69",
            "raw_upload_location": "http://minio:9000/archive/v4/repos/57E07734BA4ADEAC0CB7FE068EF0FBCA/static_analysis/files/b29af641-1ae3-302d-a652-cccd2878fb69.json?X-Amz-Algori...59f8c446ff",
            "state": "valid"
        },
        {
            "filepath": "wildwest/__init__.py",
            "file_hash": "d41d8cd9-8f00-b204-e980-0998ecf8427e",
            "raw_upload_location": "http://minio:9000/archive/v4/repos/57E07734BA4ADEAC0CB7FE068EF0FBCA/static_analysis/files/d41d8cd9-8f00-b204-e980-0998ecf8427e.json?X-Amz-Algori...76134e9348",
            "state": "created"
        },
        // ...
        {
            "filepath": "tests/test_six.py",
            "file_hash": "557e32ff-421e-ec9f-edea-ba35e876b10d",
            "raw_upload_location": "http://minio:9000/archive/v4/repos/57E07734BA4ADEAC0CB7FE068EF0FBCA/static_analysis/files/557e32ff-421e-ec9f-edea-ba35e876b10d.json?X-Amz-Algori...694e7f4ef2",
            "state": "valid"
        }
    ]
}

```

The `raw_upload_location` tells the customer where to upload the each individual file analysis (or as we call, a snapshot) to, and the `state_id` tells the state of that snapshot. Snapshots with `state == valid` were already checked as valid by the system (possibly on another analysis) and don't need to be uploaded again (but the customer can if they want to, we will have to iron details on that later)

Those serializers and views implement that. We might need some more logging in the future. But for now that will suffice.

Notice that this creation view will create a task for the worker. That task will be where we will check the uploads to GCS, and verify if they are valid. Having a valid upload means the customer won't need to upload it again in the future.