Adding CoverageDatapoint enum to ReportLine
As part of the iterative testing project, we have to find ways to associate per-test coverage to each line. Our current granularity is per-session (which itself includes many tests), and that won't work for this case.

So we are creating an entity called `CoverageDatapoint`. It has the fields:

- sessionid: the session id where that coverage was registered
- coverage: the coverage associated with it
- coverage_type: the coverage type associated with it (read: 'default', 'branch' or 'method')
- labels: the list of labels, which are strings, associated with that datapoint. Those labels are going to usually be the test names

The reason why we have datapoints each with its own coverage and labels, instead of just adding labels to `ReportLine` is that, once a particular label is removed from a Report, we want to make sure that, if the remaining labels have a different coverage than the removed one, the new coverage for that reportline has to represent the combined coverage of those labels

Each `ReportLine` will then have multiple `CoverageDatapoint`s. We were going to just add that to the ReportLine.
But there is now a problem that there would be a new extra field.

We would like to repurpose `messages`. The few places that intended to use messages originally are commented out and never got actually used. This was pre-2018 ( https://github.com/codecov/codecov.io/commit/3cbabe78ef71e4ae8b79893890f58c8aa79c574d#diff-71c497141b2eef041da62dace565fa12b3067191ff0bf6e73d603bff04ff0baaR8 ). There is no benefit in keeping this field there.

This is not going to work, though, as there are many small things using this `None` field, specially on legacy and on codecov/report