Adding raise to exception catching
A long time ago, we added some rust logic to calculate comparisons. As it was experimental, we added some `except Exception` around it, so bugs inside it wouldn't block the full notification flow (which was the piece of logic using it).

We were aware that other exceptions (like Soft Time Limit Exceptions) would also be swallowed here. But it was worth it because:
- Soft time limits mean the task would not have continued either way, so nothing was really lost
- The number of soft time limits was probably small compared to the number of errors this could throw

But it's been a while, and now all errors happening inside here are soft time limit. We should just remove the try except from here