Upgrading cassette files
This is a change caused by a rabbit hole. Something needed something that could use this

`codecov_vcr` is a system that, on the first time a test is run, saves out outgoing https and its responses in a YAML. On later tests executions, it just intercepts the calls and replays them from the saved YAML

This saves time on later executions and provides a less flaky test system (because it doesn't rely on external factors)

Seemingly latest VCR version saves data slightly different, and when something changes in an actual test change, it's hard to tell which parts of the cassette changed because of the test change and which ones changed because of the cassette format change.

So we are forcing a bunch of cassette format changes in this PR

To achieve this, we simply added `cassete_maker._save(force=True)` locally to the `codecov_vcr`, and re-ran all the tests.

Notice a bunch of new empty cassettes. Seemingly the tests use them, but they were not being stored because they are not empty. I am taking this opportunity to save them as well, because otherwise some test in CI might start making calls all of a sudden and it won't fail without a cassette file there