Upgrading shared to 0.8.3
To update the cassettes we changed `codecov_vcr` from `once` mode to `append` mode, as we know the only changes are some new API calls from Github pulls API

For the future, we might want to change this from `codecov_vcr` to `respx_vcr`, as it is easier to update

Notice how most tests gained only one HTTP call. That's the new call we are introducing to find PR requests associated with a commit.