Dealing with cases where diffs have no before
We have a `before` structure in our git diffs. It's meant to hold the name a file had before, in case of renames. In case of no renames or in case the file is brand new, we have `"before": None`.

Some bits of our code assumed that the value "before" would always be present. But as we learned, in cases like the "binary", the "before" field is not there (for whatever reason), and we can't assume it's there. But we can just change the function to use `get("before")` instead of `["before"]` and the effect will be the same