Org wide tokens/admin crud (#1121)
* ability to create and delete org tokens from django admin

We are going to leverage Django admin to allow users to create / delete / refresh orglevel tokens. That happens from the Owner details page. Only staff users are able to access that information and make changes to tokens.

Further requirements are the ability to refresh tokens (in case there's a leak and a token needs to be invalidated quickly) . Also it has been decided that at this time only 1 token per org should exist, so staff can only create a new token if an org doesn't already have one.

* add token_type and valid_until to org-wide token
* add refresh option to inline form


* make UPLOAD the default value for orgleveltoken type in Django ORM level.
This is because when searching via the org level token service, if the org had a token that was type-less (which is possible) the service would create a new token, rather the get the existing one. Like so every token is created with a default UPLOAD type, and we make sure the service gets tokens when they exist, since the only filter is the owner. It will also be better for the future when we have multiple possible token types.

* Don't auto-generate tokens for orgs in enterprise-cloud that don't have one.
This change effectively goes agains the rationale of automatically generating a orgleveltoken for all orgs that can have one. The reason for this is because of poor use of Django's signals. Specifically if you create a token for an org from the django admin page, because the Owner model is saved first it will actually create 2 tokens - the one automatically created when Owner is saved and the one you actually wanted to create. That is a big problem because the admin interface breaks and most importantly because we fail the constraint that an org should only have 1 orgleveltoken.

So now org tokens need to be manually created. That can be done from the django admin panel for the owner.

* List org level token panel for all owners

Previously only owners in the enterprise-cloud plans saw the Org Level Token form (which makes a lot of sense). The issue is that when you change an owner plan from the admin panel and saved the user it failed saying the form didn't have fields related to the org tokens and the user plan didn't change.

So I just added the form to all owners, and updated the title message saying it's only available for enterprise cloud users. I'll be checking options to make it so that only enterprise cloud users see that, but it might not be worth the effort.