Decoupling enum name from its values
We have a bug where the worker is writing `carriedforward` to a column, and API is checking for `carryforwarded`.

We are going to fix that.

But there is also the underlying problem that changing it would make a couple of thing weird: we use `enum_value.name` in a few places. And while this is mostly fine, it ends up tying the enum variable name to things like database and such. This makes it so changing enum names, which was supposed to be a code-style only change, have actual effects on the code.

So we are decoupling that a little. But to do that, we need to change `CodecovDatabaseEnum` and all its implementing classes to have a special field `db_id`.

From there, we are now adding a property `db_name` to `UploadType`. That's the value we are going to use for the db queries. So no more wrongly using the enum name, which will break things once changed.

We are also purposefully renaming most enums to uppercase to `UploadType.carriedforward`. This is the python standard anyway, and changing them will make all the tests using enums break on API and worker, forcing us to adapt them to this new flow

Just a note that most of the time using `Enum.name` is fine. It's just on this case that we having a conflict with an existing Enum `SessionType` that had different flows