Implementing Single File Analyzer
This change implements a class `SingleFileSnapshotAnalyzer` that is capable of taking the static analysis of a single file and produce calculations on top of it

The first calculations are:

1. `get_corresponding_executable_line`: Imagine a statement that is multi-line. Something like

```
10 | simple = [
11 |     x
12 |     for x in elem_list
13 |     if x > 5 and (function(x) <= 10 or x in exclude_list)
14 |     for elem_list in banana_list
]
```

If one edits the ` > 5 ` part it will be an edit on line 13. But in a lot of languages (like Python), line 13 is not a line that will register on the report. Only line 10 registers on the report.

But an edit on line 13 does affect line 10. So we need to know that when making calculations about lines affected about a certain git change

This is what this function does. And the heavy lifting was alredy done by the codecov-cli at this point. We receive a list of statements and, for each statements, which lines are involved in it.

There is `len` which tells the len of the statement (after the first line) and `extra_connected_lines` in case a line far from the line is part of that line statement. the `extra_connected_lines` is for example useful on "if elses" since the else is technically part of the if line, but it might be very far from it.

2. `get_antecessor_executable_line`

That one is trickier: We want to know, given an executable line, which executable line that antecedes is and was unaffected by a change. Sometimes it might be just the previous lines. Sometimes we might find that this affected line was the first line of a certain function, and therefore this whole function was affected by this change. Or we find that this is the first line of the file, and the whole file was affected by this change

The ultimate goal is to find something unaffected by a change, and from there, find the corresponding line on the base commit. This is tricky and might not make much sense now, but later PRs will show how it is used