Include healthcheck as a task in install yaml with interval_seconds setting (#313)
* Include healthcheck as a task in install yaml with interval_seconds setting.

Related ticket: [CODE-2099](https://codecovio.atlassian.net/browse/CODE-2099)

We're including a health check tasks tha runs by default every 10s.
To make that time configurable via the install yaml a new fields was added under setup > tasks > healthcheck.
Notice that the queue in which the healthcheck runs cannot be configured. It is a hardcoded value (check CODE-2100).

* Replace custom verifier (ensure_positive_number) by lib-based one (min)

Giving preference to lib tools we are to replace the validation of the healthcheck interval to use the `min` feature of cerberus.
Docs [here](https://docs.python-cerberus.org/en/stable/validation-rules.html#min-max).

Notice that it uses `__gt__` so `"min": 0 == > 0`, which is what we want.