Change default GITLAB_ENTERPRISE_REDIRECT_URI
When requesting access+refresh token from GLE we need to match the redirect_uri from the 1st request (yes, there are 2 of them).

One of the default values is in api (`/login/gitlab_enterprise`) the other is in shared (`/login/gle`) that causes the login to fail.

This change sets the default value in api to be the same as shared's making the login work with default values