Make healthcheck task interval configurable (#955)
* Make healthcheck task interval configurable

The beat schedule for health_check_task is now configurable under setup > tasks > healthcheck > interval_seconds.
The default value is 10.

Creation of `get_health_check_interval_seconds` is necessary because if the install YAML is invalid the configuration is used "as is" rather than throwing error so we can't make sure that the value will be indeed valid. It always falls back to the default.

Moved other healthcheck specific settings (e.g. the queue name) to `helpers.health_check`.

* Add warn log when healthcheck interval config is invalid