Fix `test_get_workflow_run` test (#318)
This integration test didn't have a cassete for it, and was using a run from a repo that is owned by someone outside the company.
Recently it started to break and is affecting the CI runs

This change changes the repo used to the django repo (which is public) and adds a cassete for it.

ps.: ideally codecov-e2e user would be used but I couldn't find the credentials to login on github
pss.: the token was not changed. But test is passing nonetheless.