Fixing Node column check
On Node we currently have the logic:

```
if not sl or sc + 1 == ec:
    return None, None, None
```

which essentially says "If there is no start line, or the end column is only one more than the the start column, ignore this statement"

Which feels reasonable, since one-char statements are probably not statements. This feel like something Steve saw in the beginning and fixed.

But as pointed out in the ticket, this `sc + 1 == ec` part of the check is missing an important thing: it only makes sense if that's on the same line. Multi-line statements might end up on a `sc + 1 == ec`  case, but it's not a one-char statement, since it spans multiple lines. The column matching is more a coincidence than not.

With all that said, there is a small chance we are missing a case that Steve saw. But as we understand it, this code needs fixing. We will have to monitor to see if how much this is happening

Also fixing tests affected by it