Remove `psycopg2` from enterprise build due to licensing issues (#989)
* Remove `psycopg2` from enterprise build due to licensing issues

Related ticket: [CODE-2289](https://codecovio.atlassian.net/browse/CODE-2289)

`psycopg2` is licensed under the LGPL license, and we can't distribute code with that license. Unfortunately for us we use it as the postgres driver with SQLAlchemy (so a prettty important package) and we technically distribute it for the enterprise release.

These changes, inspired by [this Stack Overflow question](https://stackoverflow.com/questions/47350078/importing-external-module-in-single-file-exe-created-with-pyinstaller) will allow us to not bundle `psycopg2` in the enterprise release (thus not distributing it and complying to the license) but also still having the worker... work.

First, ` --exclude-module psycopg2` in `package.sh` makes sure that PyInstaller will not include `psycopg2` in the binary it created for the enterprise release. Notice that this will make it not work out of the box, because we will not have a driver to work with SQLAlchemy.

Second, we were directly using `psycopg2` in `base.py`. To have a dynamicly linked lib it needs to be technically replaceble. That won't work if we directly depend on a lib. Fortunatelly the only place that directly used `psycopg2` was error management. By wrapping it around a function we can make sure we only import it if that is the driver being used (it's the default driver). That is why the `_analyse_error` function was created.

How do we know what dialect we're using? From the database_url. [These are the docs](https://docs.sqlalchemy.org/en/14/core/engines.html#postgresql). Notice how we can configure a specific driver. Technically worker could also work with `pg8000` after this, cause we won't be using `psycopg2` anywhere. The `resolve_driver` helper function checks that.

The last bit of code we need is the driver that will work with SQLAlchemy after all. We will be loading it externally from a folder. To make it easier for users feeling adventurous we can configure the path of that folder, and `get_dbapi_external_folder` returns the configred path to us and extends the `PYTHON_PATH` to check the configured folder.

In the end, the user will need either `psycopg2` or `pg8000` (or any other driver accepted by SQLAlchemy) installed in the folder configured under `services.database.driver_external_folder`, and will need that folder mounted on the enterprise release image for it to work.

There's another catch though: the driver needs to be installed for the same **python version** and **architecture** than the one we used to compile the enterprise image. So to help our customers we will also offer open source images with the driver pre-installed, most likely.

* Move external_deps import to `main.py` on worker setup, remove `resolve_dialect_logic`.

To avoid global-level behavior the `sys.path.append` logic that extends PYTHON_PATH to include external_deps folder is now in the `setup_worker` function in `main.py`. The default value of external folder and the function name are more generic.

Also removing the `resolve_dialect` logic because we didn't have to do that anyway. Still I wanted to make worker not-explicitly dependent on `psycopg2` so a try / except seems to be a good compromise.